delete-daemon/pyproject.toml
[tool.poetry]
name = "delete-daemon"
version = "0.1.0"
description = ""
authors = ["ov3rwrite <ilyabelykh123@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
aio-pika = "^9.4.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
minio = "^7.2.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
